// schemaregistry/confluent/type/decimal_pb.ts
// @generated by protoc-gen-es v2.12.0 with parameter "target=ts"
// @generated from file confluent/type/decimal.proto (package confluent.type, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file confluent/type/decimal.proto.
 */
export const file_confluent_type_decimal: GenFile = /*@__PURE__*/
  fileDesc("Chxjb25mbHVlbnQvdHlwZS9kZWNpbWFsLnByb3RvEg5jb25mbHVlbnQudHlwZSI6CgdEZWNpbWFsEg0KBXZhbHVlGAEgASgMEhEKCXByZWNpc2lvbhgCIAEoDRINCgVzY2FsZRgDIAEoBWIGcHJvdG8z");

/**
 * @generated from message confluent.type.Decimal
 */
export type Decimal = Message<"confluent.type.Decimal"> & {
  /**
   * The two's-complement representation of the unscaled integer value in big-endian byte order
   *
   * @generated from field: bytes value = 1;
   */
  value: Uint8Array;

  /**
   * The precision (zero indicates unlimited precision)
   *
   * @generated from field: uint32 precision = 2;
   */
  precision: number;

  /**
   * The scale
   *
   * @generated from field: int32 scale = 3;
   */
  scale: number;
};

/**
 * Describes the message confluent.type.Decimal.
 * Use `create(DecimalSchema)` to create a new message.
 */
export const DecimalSchema: GenMessage<Decimal> = /*@__PURE__*/
  messageDesc(file_confluent_type_decimal, 0);